introduce unified interface based on AbstractNFFTPlan (+ some cleanup)
